Refactor code to allow for more tests to be added.
Added better testing for actors
Improved comments on test code and actors
Refactored some methods on the actors to allow better testing
Created new bash scripts for testing which for the temp script
will return 19. Otherwise the testing needs to be done on a machine
with the thermometer installed.
Unfortunately this means we cannot test with the hardware. This will need
to be done manually.